+++ x-pack/plugins/security_solution/server/usage/endpoints/endpoint_metrics.test.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+describe('test stub', () => {
+  it('test stub one', () => {
+    expect(true).toBeTruthy();
+  });
+});
